Auto-detect max APDU length from hardware driver capabilities
Each hardware driver now reports its maximum KNX frame length:
- TPUART v1/v2/v2+: 63 bytes (64-byte send buffer, 6-bit index)
- FT12/USB: 23 bytes (conservative: standard frame only)

Router::maxFrameLength() computes the minimum across all bus
drivers. IP servers derive maxAPDULength = maxFrameLength - 8
(extended frame overhead) at startup, unless overridden by the
'max-apdu-length' config option.

This prevents ETS from sending frames larger than the hardware
can handle, which would cause corruption on the UART link.

Type widened from uint8_t to uint16_t to support APDU lengths
>255. PID_MAX_APDULENGTH response now encodes both bytes.

+++ src/libserver/eibnetserver.h
@@ -78,7 +78,7 @@ public:
   ConnType type = CT_NONE;
   int no;
   bool nat;
-  uint8_t maxAPDULength;
+  uint16_t maxAPDULength;
 
   ev::timer timeout;
   void timeout_cb(ev::timer &w, int revents);
@@ -188,7 +188,7 @@ private:
   std::string interface;
   std::string servername;
   ev::tstamp keepalive;
-  uint8_t maxAPDULength;
+  uint16_t maxAPDULength;
   IniSectionPtr router_cfg;
   IniSectionPtr tunnel_cfg;
